remove unnecessary subcrates
