+++ Cargo.toml
@@ -9,8 +9,6 @@ edition = "2021"
 members = [
 	"client/consensus/qpow",
 	"dilithium-crypto",
-	"dilithium-crypto/sign",
-	"dilithium-crypto/verify",
 	"node",
 	"pallets/balances",
 	"pallets/merkle-airdrop",
@@ -56,7 +54,6 @@ uuid = { version = "1.7.0", features = ["serde", "v4"] }
 
 # Own dependencies
 dilithium-crypto = { path = "./dilithium-crypto", default-features = false }
-hdwallet = { path = "./dilithium-crypto/hdwallet", default-features = true }
 pallet-balances = { path = "./pallets/balances", default-features = false }
 pallet-merkle-airdrop = { path = "./pallets/merkle-airdrop", default-features = false }
 pallet-mining-rewards = { path = "./pallets/mining-rewards", default-features = false }
