--- dilithium-crypto/hdwallet/Cargo.toml
@@ -1,13 +1,0 @@
-[package]
-name = "hdwallet"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-rusty-crystals-dilithium = { workspace = true, default-features = false }
-
-[features]
-default = ["std"]
-std = [
-	# Note: We don't include rusty-crystals-dilithium/std here since it doesn't have that feature
-]
