--- dilithium-crypto/hdwallet/README.md
@@ -1,5 +1,0 @@
-# HD Wallet for PQ Crystals
-
-no_std compatible
-
-TBD
