--- dilithium-crypto/sign/README.md
@@ -1,3 +1,0 @@
-TBD
-
-pq crystals signer
