--- dilithium-crypto/verify/Cargo.toml
@@ -1,8 +1,0 @@
-[package]
-name = "verify"
-version = "0.1.0"
-edition = "2024"
-
-[dependencies]
-rusty-crystals-dilithium.workspace = true
-log.workspace = true
