--- dilithium-crypto/verify/README.md
@@ -1,3 +1,0 @@
-# PQ Crystals Verifier
-
-no_std
