--- dilithium-crypto/verify/src/lib.rs
@@ -1,13 +1,0 @@
-#![no_std]
-
-use rusty_crystals_dilithium::ml_dsa_87::PublicKey;
-
-pub fn verify(pub_key: &[u8], msg: &[u8], sig: &[u8]) -> bool {
-    match PublicKey::from_bytes(pub_key) {
-        Ok(pk) => pk.verify(msg, sig, None),
-        Err(e) => {
-            log::warn!("public key failed to deserialize {:?}", e);
-            false
-        }
-    }
-}
