+++ runtime/tests/transactions/integration.rs
@@ -45,7 +45,8 @@ mod tests {
 
         // Generate a keypair
         let entropy = [0u8; 32]; // Fixed entropy of all zeros
-        let keypair = hdwallet::generate(Some(&entropy)).expect("Failed to generate keypair");
+        let keypair =
+            dilithium_crypto::generate(Some(&entropy)).expect("Failed to generate keypair");
         let pk_bytes: [u8; PUB_KEY_BYTES] = keypair.public.to_bytes();
 
         println!(
@@ -157,7 +158,8 @@ mod tests {
 
         // Generate a keypair
         let entropy = [0u8; 32]; // Fixed entropy of all zeros
-        let keypair = hdwallet::generate(Some(&entropy)).expect("Failed to generate keypair");
+        let keypair =
+            dilithium_crypto::generate(Some(&entropy)).expect("Failed to generate keypair");
         let pk_bytes: [u8; PUB_KEY_BYTES] = keypair.public.to_bytes();
         let account_id = PoseidonHasher::hash(&pk_bytes).0.into();
         let id = Address::Id(account_id);
@@ -169,7 +171,8 @@ mod tests {
 
         // Sign payload with a different key
         let entropy2 = [1u8; 32]; // Fixed entropy of all zeros
-        let keypair2 = hdwallet::generate(Some(&entropy2)).expect("Failed to generate keypair");
+        let keypair2 =
+            dilithium_crypto::generate(Some(&entropy2)).expect("Failed to generate keypair");
         let sig_bytes_wrong_key = keypair2.sign(&msg, None, false);
         let signature_wrong_key = ResonanceSignature::try_from(&sig_bytes_wrong_key[..])
             .expect("Signature length mismatch");
@@ -217,7 +220,8 @@ mod tests {
 
         // Generate a keypair
         let entropy = [0u8; 32]; // Fixed entropy of all zeros
-        let keypair = hdwallet::generate(Some(&entropy)).expect("Failed to generate keypair");
+        let keypair =
+            dilithium_crypto::generate(Some(&entropy)).expect("Failed to generate keypair");
         let pk_bytes: [u8; PUB_KEY_BYTES] = keypair.public.to_bytes();
 
         // Create and sign a payload
@@ -277,7 +281,8 @@ mod tests {
 
         // Generate a keypair
         let entropy = [0u8; 32]; // Fixed entropy of all zeros
-        let keypair = hdwallet::generate(Some(&entropy)).expect("Failed to generate keypair");
+        let keypair =
+            dilithium_crypto::generate(Some(&entropy)).expect("Failed to generate keypair");
         let pk_bytes: [u8; PUB_KEY_BYTES] = keypair.public.to_bytes();
 
         // Create and sign a payload
